+++ src/Armature.Core/src/IPatternTreeNode.cs
@@ -34,7 +34,7 @@ namespace Armature.Core
     ///   Returns all matched build actions for the <paramref name="unitSequence" />. All actions are grouped by a building stage
     ///   and coupled with a "weight of matching". See <see cref="WeightedBuildActionBag" /> for details.
     /// </returns>
-    WeightedBuildActionBag? GatherBuildActions(ArrayTail<UnitId> unitSequence, long inputWeight);
+    WeightedBuildActionBag? GatherBuildActions(ArrayTail<UnitId> unitSequence, int inputWeight);
 
 
     // IPatternTreeNode UseBuildAction(IBuildAction buildAction, object buildStage); //TODO: change parameters position
